refactor: moving InputHelper to separate file. Fix issues.
Not sure what to do with logging??
